+++ packages/frontend/src/pages/Profile/NotificationsPage/NotificationsPage.tsx
@@ -24,20 +24,35 @@ export const NotificationsPage = () => {
   const { settingsGroups, settings, settingsSearch } = useSettings({
     isLoading: isLoadingUser || isLoadingParties,
     options: {
-      excludeGroups: [SettingsType.contact, SettingsType.alerts, SettingsType.profiles, SettingsType.favorites],
+      excludeGroups: [SettingsType.contact, SettingsType.primary, SettingsType.favorites],
+      groups: {
+        [SettingsType.alerts]: {
+          title: 'Varslingsadresser',
+        },
+        [SettingsType.profiles]: {
+          title: 'Varslingsprofiler',
+        },
+        [SettingsType.persons]: {
+          title: 'Varslinger for andre personer',
+        },
+        [SettingsType.companies]: {
+          title: 'Varslinger for virksomheter',
+        },
+      },
     },
   });
 
   return (
     <PageBase>
       <Breadcrumbs items={getBreadcrumbs(t('sidebar.profile'), t('sidebar.profile.notifications'), search)} />
-      <Heading size="xl">{t('sidebar.profile.notifications')}</Heading>
+      <Heading size="xl">Varslingsinnstillinger</Heading>
       <Toolbar
         search={{
           ...settingsSearch,
-          placeholder: t('parties.search.placeholder'),
+          placeholder: 'Søk i varslinger',
         }}
       />
+      {settings.length === 0 && <Heading size="lg">{t('profile.settings.no_results')}</Heading>}
       <SettingsList items={settings} groups={settingsGroups} />
     </PageBase>
   );