+++ packages/frontend/src/pages/Profile/Settings/Settings.tsx
@@ -32,21 +32,21 @@ export const Settings = () => {
   const { settingsGroups, settings, settingsSearch } = useSettings({
     isLoading: isLoadingUser || isLoadingParties,
     options: {
-      excludeGroups: [SettingsType.alerts],
+      includeGroups: [SettingsType.contact],
     },
   });
 
   return (
     <PageBase>
       <Breadcrumbs items={getBreadcrumbs(t('sidebar.profile'), t('sidebar.profile.settings'), search)} />
-      <Heading size="xl">{t('sidebar.profile.settings')}</Heading>
+      <Heading size="xl">{t('profile.settings.personal_settings')}</Heading>
       <Toolbar
         search={{
           ...settingsSearch,
-          placeholder: t('parties.search.placeholder'),
+          placeholder: t('profile.settings.search_placeholder'),
         }}
       />
-      {settings.length === 0 && <Heading size="lg">Ingen treff</Heading>}
+      {settings.length === 0 && <Heading size="lg">{t('profile.settings.no_results')}</Heading>}
       <SettingsList items={settings} groups={settingsGroups} />
     </PageBase>
   );